% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_db.R
\name{append_db}
\alias{append_db}
\title{append_db}
\usage{
append_db(ttT, ttT_sub = NULL, which_matrix = NULL,
  which_tfbms = NULL, explicit_zeros = TRUE)
}
\arguments{
\item{ttT}{see infer_db}

\item{ttT_sub}{see infer_db}

\item{which_matrix}{see infer_db}

\item{which_tfbms}{see infer_db}

\item{explicit_zeros}{see infer_db}
}
\description{
append_db
}
