% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{append_matrix}
\alias{append_matrix}
\title{append_matrix}
\usage{
append_matrix(ttT = ttT, ttT_sub = ttT_sub, R = R)
}
\arguments{
\item{ttT}{a tidied topTable object}

\item{ttT_sub}{a tidied topTable object whose rows contain only "DE" genes}

\item{R}{a gene by binding motif count matrix}
}
\value{
a new gene wise tidy table with appended tfbm counts
}
\description{
Join TFBM data to ttT
}
