% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_attr}
\alias{check_attr}
\title{Check if object has an attribute}
\usage{
check_attr(x, attribute)
}
\arguments{
\item{x}{R object}

\item{attribute}{string}
}
\description{
Check if object has an attribute
}
