% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_connection.R
\name{close_connection}
\alias{close_connection}
\title{Clean up a Connection}
\usage{
close_connection(conn)
}
\arguments{
\item{conn}{RODBC, RJDBC or other supported data connection}
}
\description{
Disconnects from a database
}
\details{
closes a connection
}
\keyword{database}
