% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_append}
\alias{db_append}
\title{Append rows into a database table}
\usage{
db_append(df, table, db, ...)
}
\arguments{
\item{df}{data.frame}

\item{table}{character vector of an optional schema and table name}

\item{db}{database object returned by \code{\link{db_connect}}}

\item{...}{further parameters passed to \code{dbWriteTable}, eg to modify \code{row.names} or \code{append} (depends on driver)}
}
\description{
This is a wrapper around \code{\link{db_insert}} with the default parameters set to append to a table.
}
