% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_close}
\alias{db_close}
\title{Close a database connection}
\usage{
db_close(db)
}
\arguments{
\item{db}{database object returned by \code{\link{db_connect}}}
}
\description{
Close a database connection
}
\note{
To close a cached connection, call \code{db_close} on an object returned by \code{db_connect(..., cache = FALSE)}
}
\seealso{
\code{\link{db_connect}}
}
