% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{db_config}
\alias{db_config}
\title{Load DB connection parameters}
\usage{
db_config(db, db_config_path = getOption("dbr.db_config_path"))
}
\arguments{
\item{db}{database name reference}

\item{db_config_path}{file path specified by the \code{dbr.db_config_path} global option as a string or function returning a string, which defaults to the \code{db_config.yaml} file in the current working directory}
}
\value{
list of database parameters (eg \code{hostname}, \code{port}, \code{username}, \code{password} etc)
}
\description{
Load DB connection parameters from a YAML file and optionally decrypt secrets via Amazon KMS or load values from Amazon System Manager's Parameter Store.
}
\details{
The YAML file should include all connection parameters, including the Database Driver as an R object, see eg

\preformatted{dbreference:
  host: localhost
  port: 3306
  drv: !expr RMySQL::MySQL()
  username: foo
  password: bar}

KMS-encrypted secrets should be specified with the \code{aws_kms} type in the YAML file, such as:

\preformatted{dbreference:
  ...
  password: !aws_kms |
    ciphertext}

AWS Parameter Store values can be loaded by specifying the \code{aws_parameter} type in the YAML file, such as:

\preformatted{dbreference:
  ...
  password: !aws_parameter |
    /path/to/value}

Fields that should not be passed to \code{drv} (eg extra params not used for making the DB connection) should be specified with the \code{attr} type that will be added as attributes to the returned list.
}
\note{
You need to install the \code{botor} package and also have access to the related KMS key to be able to decrypt cipher-text.
}
