% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{db_config_encrypt_secret}
\alias{db_config_encrypt_secret}
\title{Encrypt a secret via KMS and prints the base64-encoded cipher-text to the console}
\usage{
db_config_encrypt_secret(secret, key)
}
\arguments{
\item{secret}{string of the actual secret to be encrypted/stored}

\item{key}{Amazon KMS key to be used for the encryption}
}
\value{
base64-encoded cipher-text
}
\description{
Encrypt a secret via KMS and prints the base64-encoded cipher-text to the console
}
\note{
You need to have access to the related KMS key to be able to decrypt cipher-text
}
\examples{
\dontrun{
encrypt_secret('secret sentence I want to store in the YAML file')
}
}
