% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Connect to a database using YAML configs}
\usage{
db_connect(db, cache, ...)
}
\arguments{
\item{db}{name of database, must be present in the file specified by \code{getOption('db_config_path')}}

\item{cache}{optional caching of the connection. If set to \code{TRUE}, the connection will be cached in the background and an all future \code{db_connect} calls will simply return that (even if called automatically from eg \code{db_query}) until the end of the R session or when caching on the \code{db} is disabled in a future \code{db_connect} call with explic \code{cache = FALSE}. See the examples for more details.}

\item{...}{extra parameters passed to the database driver, even ones overriding the default values loaded from the YAML config}
}
\description{
Connect to a database using YAML configs
}
\examples{
\dontrun{
## create new connection
optbak <- options()
options('db_config_path' = system.file('db_config.yml', package = 'dbr'))
con <- db_connect('sqlite')
str(con)
db_query('SELECT 42', 'sqlite')

## reusing the connection
str(db_connect('sqlite', cache = TRUE))
str(db_connect('sqlite'))
str(db_connect('sqlite'))
## kill cached connection
db_close(db_connect('sqlite', cache = FALSE))

## restore options
options(optbak)
}
}
\seealso{
\code{\link{db_close}} \code{\link{db_query}}
}
