% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{db_dark_theme}
\alias{db_dark_theme}
\title{Clean Dark Theme for ggplot}
\usage{
db_dark_theme(base_size = 12, base_family = "Open Sans",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{Size of text. Defaults to 11}

\item{base_family}{Font family. Defaults to Open Sans}

\item{base_line_size}{Line size}

\item{base_rect_size}{Size of rectangles}
}
\description{
Provides a clean, dark theme, aesthetic for ggplot
}
\examples{
library(datasets)
data(iris)

ggplot(iris, aes(x=Sepal.Width, y=Sepal.Length, color=Species)) +
  geom_point() +
  labs(title="Sepal Height vs. Sepal Width",
       subtitle='Iris Dataset',
       caption="The iris dataset is well-known and common amongst new R learners.") +
  scale_color_brewer(palette="Spectral") +
  db_dark_theme(base_size = 11, base_family = 'Open Sans') \%+replace\%
  theme(legend.position = 'right')
}
\seealso{
Other theme: \code{\link{db_light_theme}}
}
\concept{theme}
