% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_insert}
\alias{db_insert}
\title{Insert rows into a database table}
\usage{
db_insert(df, table, db, ...)
}
\arguments{
\item{df}{data.frame}

\item{table}{character vector of an optional schema and table name}

\item{db}{database object returned by \code{\link{db_connect}}}

\item{...}{further parameters passed to \code{dbWriteTable}, eg to modify \code{row.names} or \code{append} (depends on driver)}
}
\description{
Insert rows into a database table
}
\examples{
\dontrun{
options('db_config_path' = system.file('example_db_config.yaml', package = 'dbr'))
db_insert(mtcars, 'mtcars', 'sqlite')
db_append(mtcars, c('dm', 'mtcars'), 'sqlite')
}
}
\seealso{
\code{RMySQL::\link[RMySQL]{dbWriteTable}}, \code{RPostgreSQL::\link[RPostgreSQL]{dbReadTable-methods}}

\code{\link{db_append}}
}
