% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_table.R
\name{drop_table}
\alias{drop_table}
\title{Delete Table}
\usage{
drop_table(table_to_delete, conn)
}
\arguments{
\item{table_to_delete}{The table that you will delete. NOTE: if you are using the Research database you can only delete
tables associated with the schema you used to create conn (probably the services database, would have been in .renviron). NOTE
for Cornerstone you must specify the entire schema + table (sndbx__research.mc_temp) while for Research you only specify the table (mc_temp).}

\item{conn}{RODBC, RJDBC or other supported data connection}
}
\description{
If the specified tables exists, it will be deleted.
}
\details{
Removes a table from a database.
}
\keyword{database}
