% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{exonic1_utr1}
\alias{exonic1_utr1}
\title{Exonic and UTR TFBM}
\format{A TFBM matrix}
\source{
Brandt
}
\usage{
exonic1_utr1
}
\description{
A dataset containing ..
}
\keyword{datasets}
