% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_db.R
\name{extract_db}
\alias{extract_db}
\title{Title}
\usage{
extract_db(x, which_outcome = "B", methods = NULL)
}
\arguments{
\item{x}{blah}

\item{which_outcome}{in general the outcome variable may be chosen as logFC, B, etc}

\item{methods}{blah}
}
\value{
blah
}
\description{
Title
}
