% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-db.R
\name{get_db_mysql}
\alias{get_db_mysql}
\title{Get a mysql database connection}
\usage{
get_db_mysql(dbname, config = "~/.mysql.cnf")
}
\arguments{
\item{dbname}{the database name (should match descriptor in .cnf file)}

\item{config}{path to the .cnf file, incluidng file name. Default 
~/.mysql.cnf}
}
\value{
A database connection
}
\description{
Get a mysql database connection
}
\examples{
\dontrun{
con <- get_db_mysql("testdb")
con <- get_db_mysql("testdb", config = "~/.custom.cnf")

dbListFields(con, "testtable")
}

}
