% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_matrix}
\alias{get_matrix}
\title{get_matrix}
\usage{
get_matrix(ttT = ttT, which_matrix = which_matrix,
  which_tfbms = which_tfbms, explicit_zeros = explicit_zeros)
}
\arguments{
\item{ttT}{a tidied topTable object}

\item{which_matrix}{which gene by binding motif count matrix}

\item{which_tfbms}{named columns of "which_matrix"}

\item{explicit_zeros}{Add explicit zero counts to the tfbm matrix for genes in sampling frame, but not in tfbm matrix}
}
\value{
a tidied tfbm matrix
}
\description{
Unexported function to LOAD AND TIDY TFBM MATRIX
}
