% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tfbm.R
\name{get_tfbm_p_vals}
\alias{get_tfbm_p_vals}
\title{Title}
\usage{
get_tfbm_p_vals(rhs, of_in = of_in, which_matrix = "utr", dat,
  n_sim = 10000)
}
\arguments{
\item{rhs}{blah}

\item{of_in}{blah}

\item{which_matrix}{blah}

\item{dat}{blah}

\item{n_sim}{blah}
}
\value{
blah
}
\description{
maaaaa
}
