% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_ex_db.R
\name{inf_ex_db}
\alias{inf_ex_db}
\title{inf_ex_db}
\usage{
inf_ex_db(rhs, of_in = of_in, which_matrix = utr1, dat,
  n_sim = 10000)
}
\arguments{
\item{rhs}{the right hand side of a regression, as a character vector of variable names}

\item{of_in}{see infer_db}

\item{which_matrix}{see infer_db}

\item{dat}{an expression set object}

\item{n_sim}{see infer_db}
}
\value{
blah
}
\description{
A convenience function
}
\details{
Used to create a pipeline around infer_db.
}
