% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_extract_db.R
\name{infer_extract_db}
\alias{infer_extract_db}
\title{get_tfbm_p_vals}
\usage{
infer_extract_db(rhs, of_in = of_in, which_matrix = utr1, dat,
  n_sim = 10000)
}
\arguments{
\item{rhs}{the right hand side of a regression, as a character vector of variable names}

\item{of_in}{see infer_db}

\item{which_matrix}{see infer_db}

\item{dat}{an expression set object}

\item{n_sim}{see infer_db}
}
\value{
blah
}
\description{
Used to create a pipeline around infer_db.
}
