% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_postgresql.R
\name{init_postgresql}
\alias{init_postgresql}
\title{PostgreSQL Database Connection}
\usage{
init_postgresql(user, password, database)
}
\arguments{
\item{user}{the password}
}
\value{
an object that maintains a connection to the database that you can query with SQL
}
\description{
Maintains a connection to a local PostgreSQL database
}
\keyword{PostgreSQL,}
\keyword{SQL}
\keyword{database,}
