% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_sqlite.R
\name{init_sqlite}
\alias{init_sqlite}
\title{SQLite Database Connection}
\usage{
init_sqlite(database_file = NULL)
}
\arguments{
\item{database_file}{the full folder name of the SQLite database file}
}
\value{
an object that maintains a connection to the database that you can query with SQL
}
\description{
Maintains a connection to a local SQLite database
}
\keyword{SQL}
\keyword{SQLite,}
\keyword{database,}
