% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift.R
\name{is.redshift}
\alias{is.redshift}
\title{Checks if a database reference or connection is Redshift or not}
\usage{
is.redshift(db)
}
\arguments{
\item{db}{database reference by name or object}
}
\value{
boolean
}
\description{
Checks if a database reference or connection is Redshift or not
}
\keyword{internal}
