% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_as_of.R
\name{join_as_of}
\alias{join_as_of}
\title{Wrapper for data.table to perform a join as of}
\usage{
join_as_of(left_df, right_df, on, left_time, right_time,
  direction = "forward")
}
\arguments{
\item{left_df}{The left side of the join. Should be a data frame.}

\item{right_df}{The right side of the join. Should be a data frame.}

\item{on}{What column to join on.}

\item{left_time}{The name of the column within left_dt to join as of.}

\item{right_time}{The name of the column within right_df to join as of.}

\item{direction}{Which direction to join as of. Should be 'forward', 'backward', or 'nearest'}
}
\description{
Given I find the syntax used by DT to be quite tedious to learn,
this function abstracts the operation of performing a join
as of leveraging the performance of DT.
}
