% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{list_remove_intermediate_level_by_name}
\alias{list_remove_intermediate_level_by_name}
\title{Remove intermediate list by moving up the children by one level}
\usage{
list_remove_intermediate_level_by_name(l, n)
}
\arguments{
\item{l}{list}

\item{n}{list name of levels to be removed (but keeping children)}
}
\value{
list
}
\description{
Remove intermediate list by moving up the children by one level
}
\examples{
\dontrun{
l <- list(a = 1, b = list('removeme' = 2))
list_remove_intermediate_level_by_name(l, 'removeme')
}
}
\keyword{internal}
