% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tfbm.R
\name{load_tfbm_matrix}
\alias{load_tfbm_matrix}
\title{load_tfbm_matrix}
\usage{
load_tfbm_matrix(which_matrix)
}
\arguments{
\item{which_matrix}{one of "utr", "exonic", "exonic_utr"}
}
\value{
blah
}
\description{
See raw-data file for the code used to make the matrix
}
