% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_values.R
\name{map_values}
\alias{map_values}
\title{Maps elements using a names list}
\usage{
map_values(x, map, if_null = NULL)
}
\arguments{
\item{x}{Array}

\item{map}{Named list}

\item{if_null}{Value to use if not present in \code(map), if
if_null == NULL, then the missing values will be
omitted.}
}
\value{
array of mapped values
}
\description{
Given an array and a named list, maps values from the array
to the right hand side of the named list.
}
\examples{

x <- c(2, 2, 1, 2, 1, 3, 4)
map <- list('1' = 'one', '2' = 'two', '2' = 'three')

new_values <- map_values(x, map)

}
