% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Finds the mode of an array}
\usage{
mode(x, tie = c("first", "last", "random"), na.rm = FALSE)
}
\arguments{
\item{x}{List of elements}

\item{tie}{String of how to handle ties}

\item{na.rm}{Boolean, remove NA elements before evaluating}
}
\value{
mode of x
}
\description{
Given an array, will find the most common element, or mode.
}
\examples{

vals <- c(2, 2, 1, 1, 3, 3)

mode(vals, tie='first')
mode(vals, tie='last')
for (i in 1:10){
  print(mode(x0, tie='random'))
}
}
