% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_sql_server.R
\name{ms_sql_server}
\alias{ms_sql_server}
\title{SQL Server Database Connection}
\usage{
ms_sql_server(server, database, username = "", password = "")
}
\arguments{
\item{server}{the server that hosts the database}

\item{database}{the name of the database}

\item{username}{your username, if you do not include this R will prompt you when the object is created}

\item{password}{your password, if you do not include this R will prompt you when the object is created}
}
\value{
an object that maintains a connection to the server that you can query with SQL
}
\description{
Maintains a connection to a SQL Server database that you can query
}
\examples{

}
\keyword{SQL}
\keyword{Server,}
\keyword{database,}

