\name{pain}
\alias{pain}
\docType{data}
\title{
Pain Data
}
\description{
Based on a surve of nearly 10,000 patients in UK health clinics during 2010-2014 to assess the quality of care they received.
}
\usage{data("pain")}
\format{
  A data frame with 1318 observations on the following 3 variables.
  \describe{
    \item{\code{severity}}{Average of 4 responses, each on a 0-10 scale (11 levels). They measure patients' perception of pain severity - over the 7 days leading up to the survey - at (1) its worst, (2) at its least, (3) on average, and (4) right now.}
    \item{\code{interference}}{Average of 7 scores, each on a 0-10 scale (11 levels). These questions measure - over the 7 days leading up to the survey - the level of interference of pain in patient's life along the following dimensions: (1) general activity, (2) mood, (3) walking ability, (4) normal work (outside of home and housework), (5) relations with other people, (6) sleep and (7) enjoyment of life.}
    \item{\code{age}}{Age of respondents, in years.}
  }
}
\keyword{datasets}
