% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_data.R
\name{post_data}
\alias{post_data}
\title{Post Data to a Connection}
\usage{
post_data(df, table_name, conn, append = FALSE)
}
\arguments{
\item{df}{The dataframe to post to the database}

\item{table_name}{the name of the database table that you would like to create}

\item{conn}{RODBC or SQLiteConnection data connection}

\item{append}{Do you want to post data to an existing data table?}
}
\description{
Posts data to your connection. NOTE For RODBC data connections don't use the fully qualified name when posting data. You can only have tables to the database you have specified in the connection string.
}
\details{
Creates a database table from a dataframe
}
\keyword{database}
