% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull Data from a Connection}
\usage{
pull_data(sql_code, conn)
}
\arguments{
\item{sql_code}{SQL query that returns one data table}

\item{conn}{RODBC (Research Database) or RJDBC (Cornerstone) data connection}
}
\description{
Gets data from a database connection and returns the result as a tibble. Provides a common interface.
}
\details{
Gets data from a database connection and returns the result as a tibble
}
\keyword{database}

