% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift.R
\name{redshift}
\alias{redshift}
\title{Redshift Database Connection}
\usage{
redshift(factory, redshift_url, username = "", password = "")
}
\arguments{
\item{factory}{the redshift factory name}

\item{redshift_url}{the Amazon S3 URL where the Redshift instance is hosted}

\item{username}{your username, if you do not include this R will prompt you when the object is created}

\item{password}{your password, if you do not include this R will prompt you when the object is created}
}
\value{
an object that maintains a connection to the server that you can query with SQL
}
\description{
Maintains a connection to an Amazon Redshift database that you can query
}
\keyword{SQL}
\keyword{Server,}
\keyword{database,}

