% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{regress}
\alias{regress}
\title{Title
Estimates simple and multiple regression of gene-specific d.e.  estimates on tfbm(s), and the regression on the total number of tfbm sites, indep of which tf they belong to.}
\usage{
regress(y, X)
}
\arguments{
\item{y}{blah}

\item{X}{blah}
}
\value{
blah
}
\description{
Title
Estimates simple and multiple regression of gene-specific d.e.  estimates on tfbm(s), and the regression on the total number of tfbm sites, indep of which tf they belong to.
}
