% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_columns.R
\name{show_columns}
\alias{show_columns}
\title{Show Columns}
\usage{
show_columns(conn, table)
}
\arguments{
\item{conn}{RODBC, RJDBC or other supported data connection}

\item{table}{the table that you want to inspect}
}
\description{
Shows the columns in the table you specify
}
\details{
Quick way to see what columns are in a table
}

