% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tables.R
\name{show_tables}
\alias{show_tables}
\title{Show Database Tables}
\usage{
show_tables(conn, show_tables_matching = "", database = "")
}
\arguments{
\item{conn}{RODBC, RJDBC or other supported data connection}
}
\description{
Shows the tables in your data connection
}
\details{
Quick way to see what tables you have available to you
}
\keyword{database}

