% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{sql_chunk}
\alias{sql_chunk}
\title{Look up common SQL chunks to be reused in SQL queries}
\usage{
sql_chunk(key, ..., indent_after_linebreak = 0)
}
\arguments{
\item{key}{key defined in \code{\link{sql_chunk_files}}}

\item{...}{passed to \code{glue} for string interpolation}

\item{indent_after_linebreak}{integer for extra indent}
}
\value{
string
}
\description{
Look up common SQL chunks to be reused in SQL queries
}
\examples{
\dontrun{
sql_chunk_files(system.file('example_sql_chunks.yaml', package = 'dbr'))
sql_chunk('dbr.shinydemo.countries.count')

## pass it right away to a database
countries <- db_query(sql_chunk('dbr.shinydemo.countries.count'), 'shinydemo')

## example for a more complex query
cities <- db_query(sql_chunk('dbr.shinydemo.cities.europe'), 'shinydemo')
}
}
