% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{sql_chunk_files}
\alias{sql_chunk_files}
\title{List or update the list of SQL chunk files to be used in \code{\link{sql_chunk}}}
\usage{
sql_chunk_files(file, add = TRUE)
}
\arguments{
\item{file}{path}

\item{add}{by default, the new \code{file} will be added to the list of active SQL chunk files, but when set to \code{FALSE}, it will override the old list instead of appending}
}
\value{
character vector of the active SQL chunk files (invisibly after update)
}
\description{
List or update the list of SQL chunk files to be used in \code{\link{sql_chunk}}
}
