% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite}
\alias{sqlite}
\title{SQLite Database Connection}
\usage{
sqlite(database.file = "")
}
\arguments{
\item{database.file}{the full folder name of the SQLite database file}
}
\value{
an object that maintains a connection to the database that you can query with SQL
}
\description{
Maintains a connection to a SQLite database that you can query
}
\keyword{SQL}
\keyword{SQLite,}
\keyword{database,}

