% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_values.R
\name{values}
\alias{values}
\title{Get values from named list}
\usage{
values(x)
}
\arguments{
\item{x}{Named list}
}
\value{
Values of x as an array
}
\description{
Given a names list, returns an array of just the values.
Analagous to the names function.
}
\examples{

x <- list('apples'=2, 'oranges'=4)
values(x)

}
