% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wgsadbworkr_example}
\alias{wgsadbworkr_example}
\title{Get path to load wgsadbworkr example}
\usage{
wgsadbworkr_example(file_name)
}
\arguments{
\item{file_name}{Name of file}
}
\description{
\code{wgsadbworkrr} comes bundled with sample files in its
\code{inst/extdata} diretory. This function makes them easier to access.
Based on /code{readr::readr_example()}
}
\examples{
wgsadbworkr_example(file_name = "fr_5_config.tsv")

}
