% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift.R
\name{write_jsonlines}
\alias{write_jsonlines}
\title{Writes each row of a data frame as JSON into a file separated by line breaks as per \url{http://jsonlines.org}}
\usage{
write_jsonlines(df, file = tempfile())
}
\arguments{
\item{df}{data frame}

\item{file}{path}
}
\description{
Writes each row of a data frame as JSON into a file separated by line breaks as per \url{http://jsonlines.org}
}
\examples{
\dontrun{
t <- tempfile()
write_jsonlines(mtcars, t)
}
}
