% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_db.R
\name{append_matrix}
\alias{append_matrix}
\title{append_matrix}
\usage{
append_matrix(ttT = ttT, ttT_sub = ttT_sub, R = R)
}
\arguments{
\item{ttT}{blah}

\item{ttT_sub}{blah}

\item{R}{blah}
}
\value{
a new gene wise tidy table with appended tfbm counts
}
\description{
Join TFBM data to ttT
}
