% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{db_light_theme}
\alias{db_light_theme}
\title{Clean Light Theme for ggplot}
\usage{
db_light_theme(base_size = 12, base_family = "Open Sans",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{Size of text. Defaults to 11}

\item{labs}{Include y and x axis labels? Defaults to FALSE}
}
\description{
Provides a clean, light theme, aesthetic for ggplot
}
\examples{
df <- data.frame(
a = seq(1,10),
b = seq(1, 10)
)

ggplot(df, aes(x=a, y=b)) +
  geom_bar(stat='identity') +
  labs(title="Plot Title",
       subtitle='Plot Subtitle',
       y="y Axis Label",
       x="x Axis Label",
       caption="Plot caption") +
  db_light_theme(base_size = 11, base_family = 'Open Sans') +
  scale_fill_brewer(palette = 'Set1')
}
