% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-db.R
\name{dplyr_connect}
\alias{dplyr_connect}
\title{Get a dplyr sql src object, the mysql database}
\usage{
dplyr_connect(dbname,
  config = "/projects/topmed/variant_annotation/.variant-server.cnf")
}
\arguments{
\item{dbname}{the database name (should match descriptor in .cnf file)}

\item{config}{path to the .cnf file, including file name. Default
~/.mysql.cnf}
}
\value{
a dplyr sql src object
}
\description{
Get a dplyr sql src object, the mysql database
}
\examples{
\dontrun{
con <- dplyr_connect("testdb")
con <- dplyr_connect("testdb", config = "~/.custom.cnf")

dplyr_table <- tbl(con, "testtable")
}

}
