% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{exonic1_utr1}
\alias{exonic1_utr1}
\title{A gene by transcription factor binding motif matrix.}
\format{A gene by transcription factor binding motif matrix.}
\source{
biomart
}
\usage{
exonic1_utr1
}
\description{
Broadly, this contains binding loci counts for each transcription factor motif in each gene. In particular, it counts the number of times any part of the transcription factor binding motif overlaps with an exon AND simultaneously overlaps with the 1000 base pairs upstream of the transcription start site.
}
\keyword{datasets}
