% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{exonic1_utr1}
\alias{exonic1_utr1}
\title{Random Weather Data}
\format{A data frame of 7 rows and 3 columns
\describe{
 \item{Day}{Numeric values giving day of the week, 1 = Monday, 7 = Sunday}
 Day is a day
 Temp is a tmp
Weather is weather
}}
\source{
dunno
}
\usage{
exonic1_utr1
}
\description{
A dataset containing randomly generated weather data.
}
\keyword{datasets}
