% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_db.R
\name{extract_db}
\alias{extract_db}
\title{extract_db}
\usage{
extract_db(x, which_outcome = "logFC", methods = NULL)
}
\arguments{
\item{x}{the output of \code{\link{infer_db}}}

\item{which_outcome}{in general the outcome variable may be chosen as logFC, B, etc}

\item{methods}{depricated}
}
\value{
A tibble whose columns give tfbm p-values for various methods:

\itemize{
 \item p_par two-tailed p-value from parametric telis, returned  only if argument 'tt_sub' is given to \code{\link[dbr]{infer_db}}.
 \item par_p_over  = upper tailed one-sided p-value for p_par
 \item par_p_under = lower tailed one-sided p-value for p_par
 \item p_uni  = simple univariate regression of b on each tfbm
 \item p_cov  = multple regression of b on all tfbm's simultaniously
 \item p_npar = nonparametric telis, returned  only if infer_db() argument 'tt_sub' is provided and perm_telis = f
}
}
\description{
A TABLE OF P-VALUES FOR DIFFERENT METHODS:
}
\seealso{
\code{\link[dbr]{infer_db}}
}
