% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_db.R
\name{extract_db}
\alias{extract_db}
\title{extract_db}
\usage{
extract_db(x, which_outcome = "logFC", methods = NULL)
}
\arguments{
\item{x}{blah}

\item{which_outcome}{in general the outcome variable may be chosen as logFC, B, etc}

\item{methods}{blah}
}
\value{
blah
}
\description{
A TABLE OF P-VALUES FOR DIFFERENT METHODS:
p_uni  = SIMPLE UNIVARIATE REGRESSION OF B ON EACH TFBM
p_cov  = MULTPLE REGRESSION OF B ON ALL TFBM's SIMULTANIOUSLY
p_par  = parametric telis, RETURNED  ONLY IF ARGUMENT 'tT_sub' IS GIVEN TO "db"
p_npar = nonparametric telis, RETURNED  ONLY IF ARGUMENT 'tT_sub' IS GIVEN TO "db"
}
