% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_db.R
\name{get_matrix}
\alias{get_matrix}
\title{get_matrix}
\usage{
get_matrix(which_matrix = which_matrix, which_tfbms = which_tfbms,
  explicit_zeros = explicit_zeros)
}
\arguments{
\item{which_matrix}{blah}

\item{which_tfbms}{blah}

\item{explicit_zeros}{blah}
}
\value{
a tidied tfbm matrix
}
\description{
Unexported function to LOAD AND TIDY TFBM MATRIX
}
