% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_db.R
\name{get_telis}
\alias{get_telis}
\title{Title}
\usage{
get_telis(R = R, ttT_sub = ttT_sub, n_sim = n_sim)
}
\arguments{
\item{R}{blah}

\item{ttT_sub}{bla}

\item{n_sim}{blah}
}
\value{
blah
}
\description{
Title
}
