% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_tfbm_p_vals}
\alias{get_tfbm_p_vals}
\title{get_tfbm_p_vals}
\usage{
get_tfbm_p_vals(rhs, of_in = of_in, which_matrix = "utr", dat,
  n_sim = 10000)
}
\arguments{
\item{rhs}{blah}

\item{of_in}{see infer_db}

\item{which_matrix}{see infer_db}

\item{dat}{an expression set object}

\item{n_sim}{see infer_db}
}
\value{
blah
}
\description{
Used to create a pipeline around infer_db.
}
