% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tfbm.R
\name{infer_db}
\alias{infer_db}
\title{Title}
\usage{
infer_db(ttT, ttT_sub = NULL, which_matrix = NULL,
  which_tfbms = NULL, n_sim = 10000)
}
\arguments{
\item{ttT}{blah}

\item{ttT_sub}{blah}

\item{which_matrix}{blah}

\item{which_tfbms}{blah}

\item{n_sim}{blah}
}
\value{
blah
}
\description{
Title
}
