% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnf-parsing.R
\name{list_from_cnf}
\alias{list_from_cnf}
\title{List database connections available in a .cnf file}
\usage{
list_from_cnf(path)
}
\arguments{
\item{path}{the path to the .cnf file (incluidng the .cnf file name)}
}
\value{
A list of database connections listed in the .cnf file
}
\description{
List database connections available in a .cnf file
}
\examples{
databases <- list_from_cnf('~/.mysql.cnf')

}
