% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_update.R
\name{send_update}
\alias{send_update}
\title{Send an update to a Connection}
\usage{
send_update(sql_code, conn)
}
\arguments{
\item{sql_code}{SQL code that updates the data source}

\item{conn}{RODBC, RJDBC or other supported data connection}
}
\description{
Updates the data source. Provides a common interface. Used when no data table result is expected.
}
\details{
Updates the data connection.
}
\keyword{database}

