% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_topTable.R
\name{tidy_topTable}
\alias{tidy_topTable}
\title{tidy_topTable}
\usage{
tidy_topTable(x, of_in, ...)
}
\arguments{
\item{x}{the output of limma::eBays}

\item{of_in}{# string refering to a single collumn of design matrix by name}

\item{...}{any other argument to toptable}
}
\value{
a reformatted version of topTable
}
\description{
This simply reformats topTable's output
}
